Ensure full_pth isn't py_executable.

+++ virtualenv.py
@@ -1364,12 +1364,13 @@ def install_python(home_dir, lib_dir, inc_dir, bin_dir, site_packages, clear, sy
         if not os.path.islink(py_executable):
             for pth in required_symlinks:
                 full_pth = join(bin_dir, pth)
-                if os.path.exists(full_pth):
-                    os.unlink(full_pth)
-                if symlink:
-                    os.symlink(py_executable_base, full_pth)
-                else:
-                    copyfile(py_executable, full_pth, symlink)
+                if (not os.path.islink(full_pth) and os.readlink(full_pth) != py_executable):
+                    if os.path.exists(full_pth):
+                        os.unlink(full_pth)
+                    if symlink:
+                        os.symlink(py_executable_base, full_pth)
+                    else:
+                        copyfile(py_executable, full_pth, symlink)
 
     if is_win and ' ' in py_executable:
         # There's a bug with subprocess on Windows when using a first
